requirements.txt
# ===========================================
# AI PC Controller - Production Dependencies
# ===========================================

# AI Providers
ollama>=0.3.0
google-generativeai>=0.5.0
groq>=0.5.0
openai>=1.30.0
httpx>=0.27.0
aiohttp>=3.9.0

# GUI Framework
customtkinter>=5.2.0
Pillow>=10.3.0
pystray>=0.19.5
darkdetect>=0.8.0

# PC Control
pyautogui>=0.9.54
pygetwindow>=0.0.9
keyboard>=0.13.5
mouse>=0.7.1
psutil>=5.9.8
pycaw>=20240210; platform_system == "Windows"
comtypes>=1.3.1; platform_system == "Windows"
wmi>=1.5.1; platform_system == "Windows"
pywin32>=306; platform_system == "Windows"
screen-brightness-control>=0.22.0
monitorcontrol>=3.1.0

# Voice Control
SpeechRecognition>=3.10.4
pyttsx3>=2.90
pyaudio>=0.2.14; platform_system == "Windows"
pvporcupine>=3.0.0
webrtcvad>=2.0.10
sounddevice>=0.4.6
scipy>=1.13.0
numpy>=1.26.0

# Web Automation
selenium>=4.20.0
webdriver-manager>=4.0.1
beautifulsoup4>=4.12.3
requests>=2.31.0

# Database
SQLAlchemy>=2.0.30
alembic>=1.13.1

# Configuration
pyyaml>=6.0.1
python-dotenv>=1.0.1
pydantic>=2.7.0
pydantic-settings>=2.2.0

# Utilities
loguru>=0.7.2
apscheduler>=3.10.4
watchdog>=4.0.0
cryptography>=42.0.5
pyperclip>=1.8.2
humanize>=4.9.0
rich>=13.7.1
click>=8.1.7
tqdm>=4.66.2

# Networking
netifaces>=0.11.0
speedtest-cli>=2.1.3

# Windows Specific
winshell>=0.6; platform_system == "Windows"
pywin32-ctypes>=0.2.2; platform_system == "Windows"
